Invert pin signals when used without INPUT_PULLUP pins
The rotary state machine is geared towards being used with pins configured as INPUT_PULLUP. In order to use the same state machine with INPUT pins, the signals read from the input pins need to be inverted.

This change adds a state variable and inverts pin inputs when the rotary is used without pullup pins.